Remove DI to ShowUnusedMySQLTables command :(
The database connection cannot be injected if not configured (and zauberlehrling might be used in no database environments), and this can cause e.g. listing the commands and the Travis build to fail.

In the future, this might be handled better with lazy commands: https://symfony.com/blog/new-in-symfony-3-4-lazy-commands